done drafting

+++ examples/llm/tech_qa.py
@@ -165,7 +165,7 @@ def make_dataset(args):
         """
         # encode the raw context docs
         embedded_docs = model.encode(context_docs, output_device=device,
-                                     batch_size=sent_trans_batch_size / 4)
+                                     batch_size=int(sent_trans_batch_size / 4))
         # k for KNN
         knn_neighsample_bs = 1024
         # number of neighbors for each seed node selected by KNN
